requirements.txt
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-llms-openai-like>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-embeddings-gemini>=0.1.0
llama-index-embeddings-huggingface>=0.2.0
llama-index-readers-file>=0.1.0
google-generativeai>=0.4.0
neo4j>=5.18.0
asyncpg>=0.29.0
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.2.0
networkx>=3.0
sentence-transformers>=2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
email-validator>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
